Refactor AsJsonRpcClient from extension method to direct RunBuilder method
